Split game setup helpers to enforce 1000-line part cap.
Extract town assignment, naming, and bootstrap/diagnostic routines into dedicated setup helper part files so all part units remain below the global 1000-line policy while preserving existing behavior.

Made-with: Cursor
