+++ packages/colonizethis_logic/lib/src/setup/game_setup_helpers_bootstrap.dart
@@ -1,0 +1,308 @@
+part of 'game_setup_helpers.dart';
+
+/// Adds starting civilian units for each civilian-owning faction at its capital tile.
+Game addStartingUnits({required Game game, required GameSetupConfig config}) {
+  var oldWorldUnits = List<Unit>.from(game.worldState.oldWorld.units);
+  var newWorldUnits = List<Unit>.from(game.worldState.newWorld.units);
+
+  Iterable<
+    ({
+      String id,
+      String? capitalProvinceId,
+      CapitalTile? capitalTile,
+      bool requireCapitalTile,
+    })
+  >
+  civilianOwners() sync* {
+    for (final player in game.players) {
+      yield (
+        id: player.id,
+        capitalProvinceId: player.capitalProvinceId,
+        capitalTile: player.capitalTile,
+        requireCapitalTile: true,
+      );
+    }
+    for (final minor in game.minorNations) {
+      yield (
+        id: minor.id,
+        capitalProvinceId: minor.capitalProvinceId,
+        capitalTile: minor.capitalTile,
+        requireCapitalTile: false,
+      );
+    }
+    for (final tribe in game.tribes) {
+      yield (
+        id: tribe.id,
+        capitalProvinceId: tribe.capitalProvinceId,
+        capitalTile: tribe.capitalTile,
+        requireCapitalTile: false,
+      );
+    }
+  }
+
+  for (final owner in civilianOwners()) {
+    final ownerId = owner.id;
+    final capitalProvinceId = owner.capitalProvinceId;
+    final capitalTile = owner.capitalTile;
+    if (capitalProvinceId == null || capitalTile == null) {
+      if (!owner.requireCapitalTile) continue;
+      throw StateError(
+        'Cannot spawn starting civilians without capital tile: owner=$ownerId',
+      );
+    }
+    final capitalTileKey = capitalTile.toTileKey();
+    final tileProvinceId = Unit.provinceIdFromTileKey(capitalTileKey);
+    if (tileProvinceId == null || tileProvinceId != capitalProvinceId) {
+      throw StateError(
+        'Capital tile/province mismatch for starting civilians: '
+        'owner=$ownerId capitalProvinceId=$capitalProvinceId '
+        'capitalTileKey=$capitalTileKey',
+      );
+    }
+    final capitalRegionId = ProvinceId.regionIdFrom(capitalProvinceId);
+
+    final unitConfig = config.startingResources.startingCivilianUnits;
+    for (final entry in unitConfig.entries) {
+      final unitType = entry.key;
+      final count = entry.value;
+      for (var k = 1; k <= count; k++) {
+        final unit = Unit(
+          id: '${ownerId}_${unitType.toLowerCase()}_$k',
+          type: unitType,
+          ownerId: ownerId,
+          locationProvinceId: capitalProvinceId,
+          status: UnitStatus.idle,
+          tileKey: capitalTileKey,
+        );
+        if (capitalRegionId == kRegionOldWorld) {
+          oldWorldUnits.add(unit);
+        } else {
+          newWorldUnits.add(unit);
+        }
+      }
+    }
+  }
+
+  return game.copyWith(
+    worldState: game.worldState.copyWith(
+      oldWorld: RegionData(
+        provinces: game.worldState.oldWorld.provinces,
+        units: oldWorldUnits,
+      ),
+      newWorld: RegionData(
+        provinces: game.worldState.newWorld.provinces,
+        units: newWorldUnits,
+      ),
+    ),
+  );
+}
+
+/// Adds starting land regiments and home-fleet ships for each Great Power.
+Game addStartingMilitaryAndNaval({
+  required Game game,
+  required GameSetupConfig config,
+  required MapTopology topologyOldWorld,
+}) {
+  final starting = config.startingResources;
+  final regimentCount = starting.initialMilitaryRegiments;
+  final shipCount = starting.initialNavalShips;
+
+  if (regimentCount <= 0 && shipCount <= 0) return game;
+
+  var oldWorldUnits = List<Unit>.from(game.worldState.oldWorld.units);
+  var newWorldUnits = List<Unit>.from(game.worldState.newWorld.units);
+  var fleets = List<Fleet>.from(game.worldState.fleets);
+  var nextSeq = game.worldState.nextShipInstanceSeq;
+  final inferredStart = inferNextShipInstanceSeqFromFleets(fleets);
+  if (nextSeq < inferredStart) nextSeq = inferredStart;
+
+  for (final player in game.players) {
+    final capitalProvinceId = player.capitalProvinceId;
+    if (capitalProvinceId == null) continue;
+
+    final regionId = ProvinceId.regionIdFrom(capitalProvinceId);
+    final localProvinceId = ProvinceId.localIdFrom(capitalProvinceId);
+
+    if (regimentCount > 0) {
+      final regimentTypeId = startingRegimentTypeForPlayer(player);
+      for (var i = 0; i < regimentCount; i++) {
+        final unit = Unit(
+          id: '${player.id}_${regimentTypeId}_reg${i + 1}',
+          type: regimentTypeId,
+          ownerId: player.id,
+          locationProvinceId: capitalProvinceId,
+          status: UnitStatus.idle,
+        );
+        if (regionId == kRegionOldWorld) {
+          oldWorldUnits.add(unit);
+        } else {
+          newWorldUnits.add(unit);
+        }
+      }
+    }
+
+    if (shipCount > 0 && regionId == kRegionOldWorld) {
+      final fullProvinceId = '$regionId|$localProvinceId';
+      final homeFleetId = homeFleetIdFor(player.id);
+      final existingIndex = fleets.indexWhere((f) => f.id == homeFleetId);
+      final existingFleet = existingIndex >= 0 ? fleets[existingIndex] : null;
+      final existingShips = existingFleet?.ships ?? const <ShipInstance>[];
+      final shipTypeId = startingShipTypeForPlayer(player);
+      final (seqAfter, newInstances) = mintShipInstances(
+        nextShipInstanceSeq: nextSeq,
+        typeIds: [for (var i = 0; i < shipCount; i++) shipTypeId],
+      );
+      nextSeq = seqAfter;
+
+      final homeFleet = Fleet(
+        id: homeFleetId,
+        ownerId: player.id,
+        seaZoneId: null,
+        inPortAtProvinceId: fullProvinceId,
+        regionId: regionId,
+        ships: [...existingShips, ...newInstances],
+      );
+      if (existingFleet == null) {
+        fleets.add(homeFleet);
+      } else {
+        fleets[existingIndex] = homeFleet;
+      }
+    }
+  }
+
+  return game.copyWith(
+    worldState: game.worldState.copyWith(
+      oldWorld: RegionData(
+        provinces: game.worldState.oldWorld.provinces,
+        units: oldWorldUnits,
+      ),
+      newWorld: RegionData(
+        provinces: game.worldState.newWorld.provinces,
+        units: newWorldUnits,
+      ),
+      fleets: fleets,
+      nextShipInstanceSeq: nextSeq,
+    ),
+  );
+}
+
+String startingRegimentTypeForPlayer(Player player) {
+  const fallbackId = 'peasant_levies';
+  final stats = regimentStatsById(fallbackId);
+  if (stats != null) return stats.id;
+  return regimentCatalog.isNotEmpty ? regimentCatalog.first.id : fallbackId;
+}
+
+String startingShipTypeForPlayer(Player _) {
+  return ShipEconomyCatalog.carrack.shipTypeId;
+}
+
+void _pushUnvisitedPpNeighborsDiag(
+  String u,
+  Map<String, Set<String>> neighbours,
+  Set<String> visited,
+  List<String> stack,
+) {
+  for (final v in neighbours[u] ?? const <String>{}) {
+    if (visited.contains(v)) continue;
+    stack.add(v);
+  }
+}
+
+typedef _LmDiag = ({int size, String minProvinceId, Set<String> provinces});
+
+List<_LmDiag> _landmassesSortedDescDiag(Map<String, Set<String>> neighbours) {
+  final visited = <String>{};
+  final out = <_LmDiag>[];
+  for (final start in neighbours.keys.toList()..sort()) {
+    if (visited.contains(start)) continue;
+    final comp = <String>{};
+    final stack = <String>[start];
+    while (stack.isNotEmpty) {
+      final u = stack.removeLast();
+      if (!visited.add(u)) continue;
+      comp.add(u);
+      _pushUnvisitedPpNeighborsDiag(u, neighbours, visited, stack);
+    }
+    final minId = comp.reduce((a, b) => a.compareTo(b) < 0 ? a : b);
+    out.add((size: comp.length, minProvinceId: minId, provinces: comp));
+  }
+  out.sort((a, b) {
+    final c = b.size.compareTo(a.size);
+    if (c != 0) return c;
+    return a.minProvinceId.compareTo(b.minProvinceId);
+  });
+  return out;
+}
+
+String lockedOwAssignFailureDiagnostics({
+  required GameSetupConfig config,
+  required MapTopology topology,
+  required List<String> seaBoundIds,
+}) {
+  final ppNbr = provincePpNeighbours(topology);
+  final sizes = ppLandComponentSizesSorted(topology);
+  final partitionOk = oldWorldPartitionMatchesLockedProfile(topology);
+  final feasibilityOk = lockedOldWorldRoleFeasibilityHolds(
+    topology: topology,
+    neighbours: ppNbr,
+  );
+  final lms = _landmassesSortedDescDiag(ppNbr);
+  final lmParts = <String>[];
+  for (var i = 0; i < lms.length; i++) {
+    final lm = lms[i];
+    var sea = 0;
+    for (final p in lm.provinces) {
+      if (isProvinceSeaBound(topology, p)) sea++;
+    }
+    lmParts.add('i=$i|sz=${lm.size}|sea=$sea|min=${lm.minProvinceId}');
+  }
+  var degMin = 1 << 30;
+  var degMax = 0;
+  var degSum = 0;
+  for (final pid in ppNbr.keys) {
+    final d = ppNbr[pid]?.length ?? 0;
+    degSum += d;
+    if (d < degMin) degMin = d;
+    if (d > degMax) degMax = d;
+  }
+  final nProv = ppNbr.length;
+  final degMean = nProv == 0 ? 0.0 : degSum / nProv;
+  const cap = 16;
+  final seaSample = seaBoundIds.length <= cap
+      ? seaBoundIds.join(',')
+      : '${seaBoundIds.take(cap).join(',')}…(+${seaBoundIds.length - cap})';
+  return 'lockedOW_diag seed=${config.seed} nodes=${topology.nodes.length} '
+      'edges=${topology.edges.length} nProv=$nProv ppSizes=$sizes '
+      'partitionOk=$partitionOk feasibilityOk=$feasibilityOk '
+      'seaboundCount=${seaBoundIds.length} seaboundSample=[$seaSample] '
+      'ppDegMin=$degMin ppDegMax=$degMax ppDegMean=${degMean.toStringAsFixed(2)} '
+      'landmasses=[${lmParts.join('; ')}]';
+}
+
+String lockedNwAssignFailureDiagnostics({
+  required GameSetupConfig config,
+  required MapTopology topology,
+}) {
+  final ppNbr = provincePpNeighbours(topology);
+  final sizes = ppLandComponentSizesSorted(topology);
+  final partitionOk = newWorldPartitionMatchesLockedProfile(topology);
+  final feasibilityOk = lockedNewWorldRoleFeasibilityHolds(
+    topology: topology,
+    neighbours: ppNbr,
+  );
+  final lms = _landmassesSortedDescDiag(ppNbr);
+  final lmParts = <String>[];
+  for (var i = 0; i < lms.length; i++) {
+    final lm = lms[i];
+    var sea = 0;
+    for (final p in lm.provinces) {
+      if (isProvinceSeaBound(topology, p)) sea++;
+    }
+    lmParts.add('i=$i|sz=${lm.size}|sea=$sea|min=${lm.minProvinceId}');
+  }
+  return 'lockedNW_diag seed=${config.seed} nodes=${topology.nodes.length} '
+      'edges=${topology.edges.length} nProv=${ppNbr.length} ppSizes=$sizes '
+      'partitionOk=$partitionOk feasibilityOk=$feasibilityOk '
+      'landmasses=[${lmParts.join('; ')}]';
+}
