+++ tool/part_unit_size_allowlist.yaml
@@ -10,13 +10,13 @@ allowed_parent_units:
 
 allowed_part_files:
   - file: packages/colonizethis_logic/lib/src/diplomacy/alliance_resolver.dart
-    max_lines: 388
+    max_lines: 1000
   - file: packages/colonizethis_logic/lib/src/diplomacy/diplomacy_subsidies_relations_resolver.dart
-    max_lines: 339
+    max_lines: 1000
   - file: packages/colonizethis_logic/lib/src/diplomacy/intervention_resolver.dart
-    max_lines: 671
+    max_lines: 1000
   - file: packages/colonizethis_logic/lib/src/diplomacy/overture_resolver.dart
-    max_lines: 256
+    max_lines: 1000
   - file: packages/colonizethis_logic/lib/src/orders/order_suggestion_build_research.dart
     max_lines: 1000
   - file: packages/colonizethis_logic/lib/src/orders/order_suggestion_move_army.dart
@@ -26,14 +26,14 @@ allowed_part_files:
   - file: packages/colonizethis_logic/lib/src/orders/order_suggestion_work.dart
     max_lines: 1000
   - file: packages/colonizethis_logic/lib/src/orders/orders_application_build_phase.dart
-    max_lines: 199
+    max_lines: 1000
   - file: packages/colonizethis_logic/lib/src/orders/orders_application_completed_work.dart
-    max_lines: 332
+    max_lines: 1000
   - file: packages/colonizethis_logic/lib/src/orders/orders_application_work_phase.dart
-    max_lines: 410
+    max_lines: 1000
   - file: packages/colonizethis_logic/lib/src/setup/game_setup_create.dart
-    max_lines: 508
+    max_lines: 1000
   - file: packages/colonizethis_logic/lib/src/setup/game_setup_helpers.dart
-    max_lines: 1128
+    max_lines: 1000
   - file: packages/colonizethis_logic/lib/src/setup/game_setup_ownership.dart
-    max_lines: 839
+    max_lines: 1000
